Reorganized files into CreateMerge folder
